Darwin specific IPC encoding primitives are implemented in WebKit/Shared but used in WebKit/Platform/IPC https://bugs.webkit.org/show_bug.cgi?id=243870 rdar://problem/98551932
Reviewed by Sam Weinig.

The base IPC implementation depends on stdlib, OS libraries, WTF.
Move MachSendRight serialization code from the client, WebKit/Shared,
to WebKit/Platform/IPC platform specific serialization headers.

This way WebKit/Platform/IPC is easier to compile in isolation, for
unit test purposes.

Also move audit_token_t to Darwin specific serialization headers.

* Source/WebKit/Platform/IPC/ArgumentCoders.h:
* Source/WebKit/Platform/IPC/darwin/ArgumentCodersDarwin.h: Added.
* Source/WebKit/Platform/IPC/darwin/ArgumentCodersDarwin.mm: Added.
Add the platform specific serialization header and implementation.

(IPC::ArgumentCoder<MachSendRight>::encode):
(IPC::ArgumentCoder<MachSendRight>::decode):
* Source/WebKit/Scripts/webkit/messages.py:
(argument_coder_headers_for_type):
* Source/WebKit/Shared/WebCoreArgumentCoders.h:
* Source/WebKit/Shared/mac/WebCoreArgumentCodersMac.mm:
(IPC::ArgumentCoder<MachSendRight>::encode): Deleted.
(IPC::ArgumentCoder<MachSendRight>::decode): Deleted.
* Source/WebKit/SourcesCocoa.txt:
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/253422@main
