+++ Source/WebKit/Scripts/webkit/tests/TestWithLegacyReceiverMessageReceiver.cpp
@@ -27,6 +27,9 @@
 #include "TestWithLegacyReceiver.h"
 
 #include "ArgumentCoders.h" // NOLINT
+#if PLATFORM(MAC)
+#include "ArgumentCodersDarwin.h" // NOLINT
+#endif
 #include "Connection.h" // NOLINT
 #include "Decoder.h" // NOLINT
 #if ENABLE(DEPRECATED_FEATURE) || ENABLE(FEATURE_FOR_TESTING)
