+++ Source/WebKit/Scripts/webkit/tests/TestWithoutAttributesMessageReceiver.cpp
@@ -27,6 +27,9 @@
 #include "TestWithoutAttributes.h"
 
 #include "ArgumentCoders.h" // NOLINT
+#if PLATFORM(MAC)
+#include "ArgumentCodersDarwin.h" // NOLINT
+#endif
 #include "Connection.h" // NOLINT
 #include "Decoder.h" // NOLINT
 #if ENABLE(DEPRECATED_FEATURE) || ENABLE(FEATURE_FOR_TESTING)
