+++ Source/WebKit/SourcesCocoa.txt
@@ -101,6 +101,7 @@ Platform/IPC/cocoa/ConnectionCocoa.mm
 Platform/IPC/cocoa/DaemonConnectionCocoa.mm
 Platform/IPC/cocoa/MachMessage.cpp
 
+Platform/IPC/darwin/ArgumentCodersDarwin.mm
 Platform/IPC/darwin/IPCSemaphoreDarwin.cpp
 
 Platform/mac/MachUtilities.cpp
